requirements.txt
djangorestframework==3.11.1
gunicorn==20.0.4
xgboost==1.1.1
